refactor: Extract file_server into separate module
