+++ src/file_manager/handlers.rs
@@ -1,0 +1,8 @@
+use actix_web::{get, http::header::ContentType, HttpResponse, Responder};
+
+#[get("")]
+pub async fn home_page() -> impl Responder {
+    HttpResponse::Ok()
+        .insert_header(ContentType::html())
+        .body("<h1>Home Page</h1>")
+}
