+++ src/file_manager/templates.rs
@@ -1,0 +1,1 @@
+
