+++ src/main.rs
@@ -2,11 +2,13 @@ use actix_web;
 use configs::ServerConfigs;
 
 mod configs;
+mod file_manager;
 mod file_server;
+mod start;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
     let configs = ServerConfigs::from_cli_args();
 
-    file_server::start(configs).await
+    start::start(configs).await
 }
